Add common avoidance package for shared libraries (#334)
* Add avoidance package

* Cleanup cmake file

* Fix test paths

* Fix code style for test

* Move status enumerations to common

* Move rviz worldloader to avoidance pkg

* Get rid of mav_msgs
