+++ avoidance/CMakeLists.txt
@@ -1,0 +1,86 @@
+cmake_minimum_required(VERSION 2.8.3)
+project(avoidance)
+
+add_definitions(-std=c++11)
+
+## Find catkin macros and libraries
+find_package(catkin REQUIRED COMPONENTS
+  roscpp
+  rospy
+  dynamic_reconfigure
+  tf
+  pcl_ros
+  mavros
+  mavros_extras
+  mavros_msgs
+  mavlink
+)
+find_package(PCL 1.7 REQUIRED)
+
+
+catkin_package(
+  INCLUDE_DIRS include
+  CATKIN_DEPENDS roscpp rospy std_msgs mavros_msgs geometry_msgs sensor_msgs message_runtime tf
+  LIBRARIES avoidance
+)
+
+###########
+## Build ##
+###########
+
+## CMake Setup
+# Build in Release mode if nothing is specified
+if(NOT CMAKE_BUILD_TYPE)
+    set(CMAKE_BUILD_TYPE Release)
+endif(NOT CMAKE_BUILD_TYPE)
+
+## Specify additional locations of header files
+include_directories(
+  include
+  ${catkin_INCLUDE_DIRS}
+  ${PCL_INCLUDE_DIRS}
+  ${YAML_CPP_INCLUDE_DIR}
+)
+
+## Declare a C++ library
+set(AVOIDANCE_CPP_FILES   "src/common.cpp"
+)
+if(NOT DISABLE_SIMULATION)
+  set(AVOIDANCE_CPP_FILES "${AVOIDANCE_CPP_FILES}"
+                              "src/rviz_world_loader.cpp")
+endif()
+add_library(avoidance     "${AVOIDANCE_CPP_FILES}")
+
+
+## Add cmake target dependencies of the library
+add_dependencies(avoidance ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
+
+## Specify libraries to link a library or executable target against
+target_link_libraries(avoidance
+  ${catkin_LIBRARIES}
+  ${YAML_CPP_LIBRARIES})
+
+#############
+## Install ##
+#############
+
+install(DIRECTORY include/${PROJECT_NAME}/
+  DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
+  FILES_MATCHING PATTERN "*.h"
+  PATTERN ".svn" EXCLUDE)
+
+#############
+## Testing ##
+#############
+
+if(CATKIN_ENABLE_TESTING)
+    # Add gtest based cpp test target and link libraries
+    catkin_add_gtest(${PROJECT_NAME}-test test/test_common.cpp
+	)
+
+    if(TARGET ${PROJECT_NAME}-test)
+      target_link_libraries(${PROJECT_NAME}-test ${PROJECT_NAME}
+                                                 ${catkin_LIBRARIES}
+                                                 ${YAML_CPP_LIBRARIES})
+    endif()
+endif()
