+++ avoidance/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package>
-  <name>local_planner</name>
-  <version>0.0.0</version>
-  <description>Avoidance module doing local planning.</description>
+  <name>avoidance</name>
+  <version>0.2.0</version>
+  <description>avoidance is a ros package for computer vision based navigtation</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
@@ -13,7 +13,7 @@
   <!-- One license tag required, multiple allowed, one license per tag -->
   <!-- Commonly used license strings: -->
   <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
-  <license>TODO</license>
+  <license>BSD</license>
 
 
   <!-- Url tags are optional, but mutiple are allowed, one per tag -->
@@ -53,7 +53,6 @@
   <build_depend>mavros</build_depend>
   <build_depend>mavros_extras</build_depend>
   <build_depend>mavros_msgs</build_depend>
-  <build_depend>avoidance</build_depend>
 
   <run_depend>dynamic_reconfigure</run_depend>
   <run_depend>message_runtime</run_depend>
@@ -68,7 +67,6 @@
   <run_depend>mavros</run_depend>
   <run_depend>mavros_extras</run_depend>
   <run_depend>mavros_msgs</run_depend>
-  <run_depend>avoidance</run_depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
