+++ avoidance/src/rviz_world_loader.cpp
@@ -1,4 +1,4 @@
-#include "local_planner/rviz_world_loader.h"
+#include "avoidance/rviz_world_loader.h"
 
 #include <ros/console.h>
