+++ avoidance/test/test_common.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include <limits>
-#include "../include/local_planner/common.h"
 #include "../include/local_planner/histogram.h"
+#include "avoidance/common.h"
 
 using namespace avoidance;
