+++ global_planner/package.xml
@@ -54,6 +54,7 @@
   <build_depend>pcl_ros</build_depend>
   <build_depend>mavros</build_depend>
   <build_depend>mavros_extras</build_depend>
+  <build_depend>avoidance</build_depend>
 
 
   <run_depend>dynamic_reconfigure</run_depend>
@@ -69,6 +70,7 @@
   <run_depend>pcl_ros</run_depend>
   <run_depend>mavros</run_depend>
   <run_depend>mavros_extras</run_depend>
+  <run_depend>avoidance</run_depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
