+++ local_planner/src/nodes/star_planner.cpp
@@ -1,6 +1,6 @@
 #include "local_planner/star_planner.h"
 
-#include "local_planner/common.h"
+#include "avoidance/common.h"
 #include "local_planner/planner_functions.h"
 #include "local_planner/tree_node.h"
